COM-897 Refactor: split and spread the data generator class in smaller specialized classes The class TestDataGenerator is growing in size and complexity; this commit breaks it down in several smaller and specialized classes which makes the code more maintainable and readable
