+++ report-testing-framework/report-testing/src/test/java/org/jembi/bahmni/report_testing/mysql_function_tests/PatientGenderIsTests.java
@@ -10,7 +10,7 @@ public class PatientGenderIsTests extends BaseReportTest {
 	@Test
 	public void shouldReturnOneWhenPatientIsMale() throws Exception {
 		// Prepare
-		int personId = testDataGenerator.createPerson(GenderEnum.MALE, new LocalDate(2000, 9, 1), "Leon", "Aziza");
+		int personId = testDataGenerator.registration.createPerson(GenderEnum.MALE, new LocalDate(2000, 9, 1), "Leon", "Aziza");
 
 		// Execute
 		String query = "SELECT patientGenderIs(" + personId + ", '" + GenderEnum.MALE + "')";
@@ -23,7 +23,7 @@ public class PatientGenderIsTests extends BaseReportTest {
 	@Test
 	public void shouldReturnOneWhenPatientIsFemale() throws Exception {
 		// Prepare
-		int personId = testDataGenerator.createPerson(GenderEnum.FEMALE, new LocalDate(2000, 9, 1), "Louise", "Palaku");
+		int personId = testDataGenerator.registration.createPerson(GenderEnum.FEMALE, new LocalDate(2000, 9, 1), "Louise", "Palaku");
 
 		// Execute
 		String query = "SELECT patientGenderIs(" + personId + ", '" + GenderEnum.FEMALE + "')";
@@ -36,7 +36,7 @@ public class PatientGenderIsTests extends BaseReportTest {
 	@Test
 	public void shouldReturnZeroWhenPatientIsNotMale() throws Exception {
 		// Prepare
-		int personId = testDataGenerator.createPerson(GenderEnum.FEMALE, new LocalDate(2000, 9, 1), "Jeff", "Wena");
+		int personId = testDataGenerator.registration.createPerson(GenderEnum.FEMALE, new LocalDate(2000, 9, 1), "Jeff", "Wena");
 
 		// Execute
 		String query = "SELECT patientGenderIs(" + personId + ", '" + GenderEnum.MALE + "')";
@@ -49,7 +49,7 @@ public class PatientGenderIsTests extends BaseReportTest {
 	@Test
 	public void shouldReturnZeroWhenPatientIsNotFemale() throws Exception {
 		// Prepare
-		int personId = testDataGenerator.createPerson(GenderEnum.MALE, new LocalDate(2000, 9, 1), "Chris", "Kutamina");
+		int personId = testDataGenerator.registration.createPerson(GenderEnum.MALE, new LocalDate(2000, 9, 1), "Chris", "Kutamina");
 
 		// Execute
 		String query = "SELECT patientGenderIs(" + personId + ", '" + GenderEnum.FEMALE + "')";
